Add stack allocation, ctor argument type checking
* Ctor's weren't really type checked, the argument were ignored. Calls
to C'tor must reference a c'tor with the correct argument types
* Added ability to create mutliple c'tors. C'tors aren't added to the
symbol table mappings, they're just appended to the class mapping,
because of the special way they're called and because you can't
reference them directly, you can only reference them by using the Class
name. We also don't want the C'tor name to shadow the class name
* C'tors with different but ambiguous types are not allowed (ie, one
takes an int and one takes an int64)
* Added syntax to stack allocate a class using a c'tor.
* CodeGen still not implemented
* Default c'tor is created for you if you don't create one
* Initializer list added, only for value types. Default initializer
c'tors are called when not specified
* Can't assign value/reference types of classes to each other.
* Includes line,column information for compile errors, finally.
Implemented on thread_local stack and passed in by grammar.