Making libs step 1: symbol table serialization
In order to import Ruddy libs, we will need to have a way to
serialize/deserialize a symbol table reprsenting a lib for later use.

Also includes ability to build multiple rdy files on the command line.

Includes a lot of refactoring, since the AST had leaked into the symbol
table.

Adds dependency to boost.