Heap Allocation - part 1
This change has heap allocation, but GC is not enabled yet. It also
moves the project to x64, since the gc util passes we are using are only
supported on x64.

At this point, the runtime is capable of finding and parsing a
statepoint, but gc isn't actually enabled yet and we're currently not
generating it. Heap allocation is a simple malloc call, without any
cleanup ever (it just keeps allocating).

The changes are getting large enough that we need a commit before
investing more in the gc side of things.

A major issue to be addressed in the next change is how to handle
address spaces. LLVM has lowering passes for handling gc-managed heap
pointers, but that means we have to mark those pointers as belonging to
that address space (1, as opposed to 0, the stack address space).
There's an open question on how to handle methods, since we silently
pass "this" as the first parameter, and since Ruddy allows classes to be
by value or by reference (a stack allocated class would be in a
different address space). Ruddy is type safe, and nowhere else would you
be able to interchangably pass reference or value types as parameters.
This has further implications for what "this" means, when we add a
"this" literal.

The answer is probably some kind of boxing, maybe via a function pointer
or something. But there's still a question on the language design. Since
Ruddy treats value and reference types as different types, where can you
pass the "this" pointer?