Making 0 the address space for everything
I attempted here to make the default addrspace (0) the one used by the
gc stuff. I did this by porting some of the gc passes over and modifying
them. Unfortunately, the gc addrspace == 1 thing is so ingrained
elsewhere in LLVM that it was proving impossible to port all that over.

The hope at the time was that we would lower all pointers, even stack
pointers, and use simple heap-bounds checking at runtime to decide
whether or not to bother with that pointer.