Unsafe Arrays
Add support for unsafe arrays and unsafe contexts. An unsafe array can
be dereferenced with no bounds checking, and is implemented as a llvm
native array type.

Also fixes the grammar so that we can deref things like an
array-of-objects index and functions that return objects.

Makes Functions possible references. This was mostly for the grammar
above to work out, but will help nicely in the future when we implement
function pointers.

Moves exception ctor implementation to cpp for faster compile times.

Has some todos. I'm leaving on vacation and want to commit, even if it's
unfinished.